i2c improvements

+++ soft/firmware/src/sensors/sensors.h
@@ -32,7 +32,7 @@ struct CompensatedData{
   int16_t   baro_altitude;      /* высота по барометру bmp085. Целое (dm)*/
   uint32_t  baro_filtered;      /* прошедшие через фильтр значение. Паскали */
   uint16_t  air_speed;          /* воздушная скорость. Фиксированная точка. (mm/s)*/
-  int8_t    temp_onboard;       /* температура c tmp75. Целые градусы. */
+  int16_t   temp_onboard;       /* температура c tmp75. 0.01 градуса. */
   // напряжение и ток бортовой сети
   uint16_t  main_voltage;      // mV
   uint32_t  main_current;      // mA
@@ -86,7 +86,7 @@ struct RawData{
   uint16_t main_voltage;
   uint16_t main_current;
   uint16_t secondary_voltage; // напряга с вторичного источника питания для серв и приемника
-  // температура c tmp75 (сотые доли градуса)
+  // температура c tmp75
   int16_t temp_tmp75;
   // температура с барометра bmp085 (десятые доли градуса)
   int16_t temp_bmp085;
